move soundcloud model and collection into shared
