+++ app/modules/playlists/collections/playlists.collection.ts
@@ -1,6 +1,6 @@
 import {Playlist} from '../models/playlist.model';
 import {Injectable} from '@angular/core';
-import {SoundcloudCollection} from '../../main/collections/soundcloud.collection';
+import {SoundcloudCollection} from '../../shared/collections/soundcloud.collection';
 
 @Injectable()
 export class Playlists<TModel extends Playlist> extends SoundcloudCollection<TModel> {
