+++ app/modules/session/collections/authenticated_user_liked_tracks.collection.ts
@@ -1,11 +1,9 @@
 import {Injectable} from '@angular/core';
-import {authenticated} from '../decorators/authenticated.decorator';
-import {SoundcloudCollection} from '../../main/collections/soundcloud.collection';
 import {Track} from '../../tracks/models/track.model';
 import {AuthenticatedUserLikedTrack} from '../models/authenticated_user_liked_track.model';
+import {SoundcloudCollection} from '../../shared/collections/soundcloud.collection';
 
 @Injectable()
-@authenticated
 export class AuthenticatedUserLikedTracks<TModel extends Track> extends SoundcloudCollection<TModel> {
   endpoint = '/me/favorites';
   model = AuthenticatedUserLikedTrack;
