+++ app/modules/shared/models/soundcloud-image.model.ts
