+++ app/modules/tracks/collections/tracks.collection.ts
@@ -1,6 +1,6 @@
 import {Track} from '../models/track.model';
 import {Injectable} from '@angular/core';
-import {SoundcloudCollection} from '../../main/collections/soundcloud.collection';
+import {SoundcloudCollection} from '../../shared/collections/soundcloud.collection';
 
 @Injectable()
 export class Tracks<TModel extends Track> extends SoundcloudCollection<TModel> {
