+++ app/modules/users/models/user.model.ts
@@ -1,6 +1,6 @@
-import {SoundcloudModel} from '../../main/models/soundcloud.model';
 import {Injectable} from '@angular/core';
-import {SoundcloudImageModel} from '../../main/models/soundcloud-image.model';
+import {SoundcloudModel} from '../../shared/models/soundcloud.model';
+import {SoundcloudImageModel} from '../../shared/models/soundcloud-image.model';
 
 @Injectable()
 export class User extends SoundcloudModel {
